Moved bulk cancellation method to separate service
