+++ src/modules/Elsa.ProtoActor/Services/ProtoActorWorkflowRuntime.cs
@@ -265,37 +265,6 @@ internal class ProtoActorWorkflowRuntime : IWorkflowRuntime
         return new CancellationResult(result?.Result ?? false);
     }
 
-    public async Task<int> CancelWorkflowAsync(IEnumerable<string> workflowInstanceIds, CancellationToken cancellationToken = default)
-    {
-        var tasks = workflowInstanceIds.Select(id => CancelWorkflowAsync(id, cancellationToken)).ToList();
-        await Task.WhenAll(tasks);
-        return tasks.Count(t => t.Result.Result);
-    }
-
-    public async Task<int> CancelWorkflowByDefinitionVersionAsync(string definitionVersionId, CancellationToken cancellationToken = default)
-    {
-        var filter = new WorkflowInstanceFilter
-        {
-            DefinitionVersionId = definitionVersionId,
-            WorkflowStatus = WorkflowStatus.Running
-        };
-        var instances = (await _workflowInstanceStore.FindManyAsync(filter, cancellationToken)).ToList();
-        await Task.WhenAll(instances.Select(instance => CancelWorkflowAsync(instance.Id, cancellationToken)));
-        
-        return instances.Count;
-    }
-
-    public async Task<int> CancelWorkflowByDefinitionAsync(string definitionId, VersionOptions versionOptions, CancellationToken cancellationToken = default)
-    {
-        // Shouldn't we get possible multiple definitions here?
-        var workflowDefinition = await _workflowDefinitionService.FindAsync(definitionId, versionOptions, cancellationToken);
-        if (workflowDefinition is null)
-            return 0;
-
-        var result = await CancelWorkflowByDefinitionVersionAsync(workflowDefinition.Id, cancellationToken);
-        return result;
-    }
-
     /// <inheritdoc />
     public async Task<IEnumerable<WorkflowMatch>> FindWorkflowsAsync(WorkflowsFilter filter, CancellationToken cancellationToken = default)
     {
